+++ R/prints.R
@@ -11,38 +11,86 @@
 #'
 #' @export
 print.design <- function(x, ...) {
+  a = x$overallResults;
   s = x$byStageResults;
-  t = x$overallResults;
-  k = length(s$informationRates)
+  k = a$kMax;
+  
   if (k>1) {
-    df = t(data.frame(s$informationRates,
-                      s$efficacyBounds,
-                      s$futilityBounds,
-                      s$cumulativeRejection,
-                      s$cumulativeFutility,
-                      s$cumulativeAlphaSpent,
-                      s$efficacyP,
-                      s$futilityP,
-                      t$overallReject,
-                      t$alpha,
-                      t$drift,
-                      t$inflationFactor))
-    df[seq(9,12), -1] <- NA # only show overall      
-    
-    
-    colnames(df) <- paste("stage", seq_len(ncol(df)), sep=" ")
+    str1 = paste0("Group-sequential trial with ", k, " stages")
   } else {
+    str1 = "Fixed design"
+  }
+  
+  
+  str2 <- paste0("Overall power: ",
+                 round(a$overallReject, 3), ", ",
+                 "overall significance level (1-sided): ",
+                 round(a$alpha, 4))
+  
+  str3 <- paste0("Drift parameter: ", round(a$drift, 3))
+  str4 <- paste0("Inflation factor: ", round(a$inflationFactor, 3))
+  
+  
+  df1 = data.frame(x = rep("", 5))
+  colnames(df1) = NULL
+  row.names(df1) = c(str1, str2, str3, str4, "")
+  
+  
+  if (k>1) {
+    b <- s[, c("informationRates", "efficacyBounds", "futilityBounds",
+               "cumulativeRejection", "cumulativeFutility",
+               "cumulativeAlphaSpent", 
+               "efficacyP", "futilityP")]
     
-    df = t(data.frame(t$overallReject,
-                      t$alpha,
-                      s$efficacyBounds,
-                      s$efficacyP,
-                      t$drift))
+    # format number of digits after decimal for each column
+    j3 <- c(1,2,3,4,5)
+    j4 <- c(6,7,8)
+    
+    b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+    b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)
+    
+    if (x$settings$typeBetaSpending != 'none') {
+      df = t(b)
+      rownames(df) = c("Information rate",
+                       "Efficacy boundary (Z-scale)",
+                       "Futility boundary (Z-scale)",
+                       "Cumulative rejection",
+                       "Cumulative futility",
+                       "Cumulative alpha spent",
+                       "Efficacy boundary (p-scale)",
+                       "Futility boundary (p-scale)")
+      
+    } else {
+      df = t(b[,c(1,2,4,6,7)])
+      rownames(df) = c("Information rate",
+                       "Efficacy boundary (Z-scale)",
+                       "Cumulative rejection",
+                       "Cumulative alpha spent",
+                       "Efficacy boundary (p-scale)")
+    }
+    
+    
+    colnames(df) <- paste("Stage", seq_len(ncol(df)), sep=" ")
+  } else {
+    b <- s[, c("efficacyBounds", "efficacyP")]
+    
+    # format number of digits after decimal for each column
+    j3 <- 1
+    j4 <- 2
+    
+    b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+    b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)      
+    
+    df = t(b)
+    
+    rownames(df) = c("Efficacy boundary (Z-scale)",
+                     "Efficacy boundary (p-scale)")
     
     colnames(df) <- NA
   }
-  rownames(df) <- sub("^[[:alpha:]][[:alnum:]]*.", "", rownames(df))
-  print( round(df,3), ..., na.print = "" , quote = FALSE )
+  
+  print(df1, ..., na.print = "" , quote = FALSE )
+  print(df, ..., na.print = "" , quote = FALSE )
   invisible(x)
 }
 
@@ -59,115 +107,224 @@ print.design <- function(x, ...) {
 #'
 #' @export
 print.lrpower <- function(x, ...) {
+  a = x$overallResults;
   s = x$byStageResults;
-  t = x$overallResults;
-  k = length(s$informationRates)
+  k = a$kMax;
+  
   if (k>1) {
-    if (t$estimateHazardRatio) {
-      df = t(data.frame(s$informationRates,
-                        s$efficacyBounds,
-                        s$futilityBounds,
-                        s$cumulativeRejection,
-                        s$cumulativeFutility,
-                        s$cumulativeAlphaSpent,
-                        s$numberOfEvents,
-                        s$numberOfDropouts,
-                        s$numberOfSubjects,
-                        s$analysisTime,
-                        s$efficacyHR,
-                        s$futilityHR,
-                        s$efficacyP,
-                        s$futilityP,
-                        s$information,
-                        s$HR,
-                        t$overallReject,
-                        t$alpha,
-                        t$numberOfEvents,
-                        t$expectedNumberOfEvents,
-                        t$numberOfDropouts,
-                        t$expectedNumberOfDropouts,
-                        t$numberOfSubjects,
-                        t$expectedNumberOfSubjects,
-                        t$studyDuration,
-                        t$expectedStudyDuration,
-                        t$accrualDuration,
-                        t$followupTime,
-                        t$fixedFollowup,
-                        t$rho1,
-                        t$rho2))
-      df[seq(17,31), -1] <- NA # only show overall      
+    str1 = paste0("Group-sequential trial with ", k, " stages")
+  } else {
+    str1 = "Fixed design"
+  }
+  
+  if (a$rho1 != 0 || a$rho2 != 0) {
+    str1 = paste0(str1, ", FH(", a$rho1, ", ", a$rho2, ")")
+  }
+  
+  
+  
+  str2 <- paste0("Overall power: ",
+                 round(a$overallReject, 3), ", ",
+                 "overall significance level (1-sided): ",
+                 round(a$alpha, 4))
+  
+  if (k>1) {
+    str3 <- paste0("Maximum # events: ",
+                   round(a$numberOfEvents, 1), ", ",
+                   "expected # events: ",
+                   round(a$expectedNumberOfEvents, 1))
+    
+    str4 <- paste0("Maximum # dropouts: ",
+                   round(a$numberOfDropouts, 1), ", ",
+                   "expected # dropouts: ",
+                   round(a$expectedNumberOfDropouts, 1))
+    
+    str5 <- paste0("Maximum # subjects: ",
+                   round(a$numberOfSubjects, 1), ", ",
+                   "expected # subjects: ",
+                   round(a$expectedNumberOfSubjects, 1))
+    
+    str6 <- paste0("Total study duration: ",
+                   round(a$studyDuration, 1), ", ",
+                   "expected study duration: ",
+                   round(a$expectedStudyDuration, 1))
+    
+  } else {
+    str3 <- paste0("Number of events: ",
+                   round(a$numberOfEvents, 1))
+    
+    str4 <- paste0("Number of dropouts: ",
+                   round(a$numberOfDropouts, 1))
+    
+    str5 <- paste0("Number of subjects: ",
+                   round(a$numberOfSubjects, 1))
+    
+    str6 <- paste0("Study duration: ",
+                   round(a$studyDuration, 1))
+  }
+  
+  str7 <- paste0("Accrual duration: ",
+                 round(a$accrualDuration, 1), ", ",
+                 "follow-up duration: ",
+                 round(a$followupTime, 1), ", ",
+                 "fixed follow-up: ", a$fixedFollowup)
+  
+  df1 = data.frame(x = rep("", 8))
+  colnames(df1) = NULL
+  row.names(df1) = c(str1, str2, str3, str4, str5, str6, str7, "")
+  
+  
+  if (k>1) {
+    if (a$estimateHazardRatio) {
+      b <- s[, c("informationRates", "efficacyBounds", "futilityBounds",
+                 "cumulativeRejection", "cumulativeFutility",
+                 "cumulativeAlphaSpent", 
+                 "numberOfEvents", "numberOfDropouts", "numberOfSubjects", 
+                 "analysisTime", "efficacyHR", "futilityHR", 
+                 "efficacyP", "futilityP", "information", "HR")]
+      
+      # format number of digits after decimal for each column
+      j1 <- c(7,8,9,10)
+      j2 <- 15
+      j3 <- c(1,2,3,4,5,11,12,16)
+      j4 <- c(6,13,14)
+      
+      b[j1] <- lapply(b[j1], formatC, format = "f", digits = 1)
+      b[j2] <- lapply(b[j2], formatC, format = "f", digits = 2)
+      b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+      b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)
+      
+      if (x$settings$typeBetaSpending != 'none') {
+        df = t(b)
+        rownames(df) = c("Information rate",
+                         "Efficacy boundary (Z-scale)",
+                         "Futility boundary (Z-scale)",
+                         "Cumulative rejection",
+                         "Cumulative futility",
+                         "Cumulative alpha spent",
+                         "Number of events",
+                         "Number of dropouts",
+                         "Number of subjects",
+                         "Analysis time",
+                         "Efficacy boundary (HR-scale)",
+                         "Futility boundary (HR-scale)",
+                         "Efficacy boundary (p-scale)",
+                         "Futility boundary (p-scale)",
+                         "Information",
+                         "HR")
+        
+      } else {
+        df = t(b[,c(1,2,4,6,7,8,9,10,11,13,15,16)])
+        rownames(df) = c("Information rate",
+                         "Efficacy boundary (Z-scale)",
+                         "Cumulative rejection",
+                         "Cumulative alpha spent",
+                         "Number of events",
+                         "Number of dropouts",
+                         "Number of subjects",
+                         "Analysis time",
+                         "Efficacy boundary (HR-scale)",
+                         "Efficacy boundary (p-scale)",
+                         "Information",
+                         "HR")
+      }
+      
     } else {
-      df = t(data.frame(s$informationRates,
-                        s$efficacyBounds,
-                        s$futilityBounds,
-                        s$cumulativeRejection,
-                        s$cumulativeFutility,
-                        s$cumulativeAlphaSpent,
-                        s$numberOfEvents,
-                        s$numberOfDropouts,
-                        s$numberOfSubjects,
-                        s$analysisTime,
-                        s$efficacyP,
-                        s$futilityP,
-                        s$information,
-                        t$overallReject,
-                        t$alpha,
-                        t$numberOfEvents,
-                        t$expectedNumberOfEvents,
-                        t$numberOfDropouts,
-                        t$expectedNumberOfDropouts,
-                        t$numberOfSubjects,
-                        t$expectedNumberOfSubjects,
-                        t$studyDuration,
-                        t$expectedStudyDuration,
-                        t$accrualDuration,
-                        t$followupTime,
-                        t$fixedFollowup,
-                        t$rho1,
-                        t$rho2))
-      df[seq(14,28), -1] <- NA # only show overall      
+      b <- s[, c("informationRates", "efficacyBounds", "futilityBounds",
+                 "cumulativeRejection", "cumulativeFutility",
+                 "cumulativeAlphaSpent", 
+                 "numberOfEvents", "numberOfDropouts", "numberOfSubjects", 
+                 "analysisTime", "efficacyP", "futilityP", "information")]
+      
+      # format number of digits after decimal for each column
+      j1 <- c(7,8,9,10)
+      j2 <- 13
+      j3 <- c(1,2,3,4,5)
+      j4 <- c(6,11,12)
+      
+      b[j1] <- lapply(b[j1], formatC, format = "f", digits = 1)
+      b[j2] <- lapply(b[j2], formatC, format = "f", digits = 2)
+      b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+      b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)
+      
+      if (x$settings$typeBetaSpending != 'none') {
+        df = t(b)
+        rownames(df) = c("Information rate",
+                         "Efficacy boundary (Z-scale)",
+                         "Futility boundary (Z-scale)",
+                         "Cumulative rejection",
+                         "Cumulative futility",
+                         "Cumulative alpha spent",
+                         "Number of events",
+                         "Number of dropouts",
+                         "Number of subjects",
+                         "Analysis time",
+                         "Efficacy boundary (p-scale)",
+                         "Futility boundary (p-scale)",
+                         "Information")
+        
+      } else {
+        df = t(b[,c(1,2,4,6,7,8,9,10,11,13)])
+        rownames(df) = c("Information rate",
+                         "Efficacy boundary (Z-scale)",
+                         "Cumulative rejection",
+                         "Cumulative alpha spent",
+                         "Number of events",
+                         "Number of dropouts",
+                         "Number of subjects",
+                         "Analysis time",
+                         "Efficacy boundary (p-scale)",
+                         "Information")
+      }
+      
     }
     
-    colnames(df) <- paste("stage", seq_len(ncol(df)), sep=" ")
+    colnames(df) <- paste("Stage", seq_len(ncol(df)), sep=" ")
   } else {
-    if (t$estimateHazardRatio) {
-      df = t(data.frame(t$overallReject,
-                        t$alpha,
-                        t$numberOfEvents,
-                        t$numberOfDropouts,
-                        t$numberOfSubjects,
-                        t$studyDuration,
-                        t$accrualDuration,
-                        t$followupTime,
-                        t$fixedFollowup,
-                        t$rho1,
-                        t$rho2,
-                        s$efficacyBounds,
-                        s$efficacyHR,
-                        s$efficacyP,
-                        s$information,
-                        s$HR))
+    if (a$estimateHazardRatio) {
+      b <- s[, c("efficacyBounds", "efficacyHR", "efficacyP", 
+                 "information", "HR")]
+      
+      # format number of digits after decimal for each column
+      j2 <- 4
+      j3 <- c(1,2,5)
+      j4 <- 3
+      
+      b[j2] <- lapply(b[j2], formatC, format = "f", digits = 2)
+      b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+      b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)      
+      
+      df = t(b)
+      
+      rownames(df) = c("Efficacy boundary (Z-scale)",
+                       "Efficacy boundary (HR-scale)",
+                       "Efficacy boundary (p-scale)",
+                       "Information",
+                       "HR")
     } else {
-      df = t(data.frame(t$overallReject,
-                        t$alpha,
-                        t$numberOfEvents,
-                        t$numberOfDropouts,
-                        t$numberOfSubjects,
-                        t$studyDuration,
-                        t$accrualDuration,
-                        t$followupTime,
-                        t$fixedFollowup,
-                        t$rho1,
-                        t$rho2,
-                        s$efficacyBounds,
-                        s$efficacyP,
-                        s$information))
+      b <- s[, c("efficacyBounds", "efficacyP", "information")]
+      
+      # format number of digits after decimal for each column
+      j2 <- 3
+      j3 <- 1
+      j4 <- 2
+      
+      b[j2] <- lapply(b[j2], formatC, format = "f", digits = 2)
+      b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+      b[j4] <- lapply(b[j4], formatC, format = "f", digits = 4)      
+      
+      df = t(b)
+      
+      rownames(df) = c("Efficacy boundary (Z-scale)",
+                       "Efficacy boundary (p-scale)",
+                       "Information")
     }
     
     colnames(df) <- NA
   }
-  rownames(df) <- sub("^[[:alpha:]][[:alnum:]]*.", "", rownames(df))
-  print( round(df,3), ..., na.print = "" , quote = FALSE )
+  
+  print(df1, ..., na.print = "" , quote = FALSE )
+  print(df, ..., na.print = "" , quote = FALSE )
   invisible(x)
 }
 
@@ -184,31 +341,75 @@ print.lrpower <- function(x, ...) {
 #'
 #' @export
 print.lrsim <- function(x, ...) {
-  s = x$overview
-  k = length(s$numberOfEvents)
+  a = x$overview;
+  k = a$kMax;
+  
   if (k>1) {
-    df = t(data.frame(s$cumulativeRejection,
-                      s$cumulativeFutility,
-                      s$numberOfEvents,
-                      s$numberOfDropouts,
-                      s$numberOfSubjects,
-                      s$analysisTime,
-                      s$overallReject,
-                      s$expectedNumberOfEvents,
-                      s$expectedNumberOfDropouts,
-                      s$expectedNumberOfSubjects,
-                      s$expectedStudyDuration))
-    df[c(7,8,9,10,11), -1] <- NA # only show overall
-    colnames(df) <- paste("stage", seq_len(ncol(df)), sep=" ")
+    str1 = paste0("Group-sequential trial with ", k, " stages")
   } else {
-    df = t(data.frame(s$overallReject,
-                      s$expectedNumberOfEvents,
-                      s$expectedNumberOfDropouts,
-                      s$expectedNumberOfSubjects,
-                      s$expectedStudyDuration))
-    colnames(df) <- NA
+    str1 = "Fixed design"
   }
-  rownames(df) <- sub("^[[:alpha:]][[:alnum:]]*.", "", rownames(df))
-  print( round(df,3), ..., na.print = "" , quote = FALSE )
+  
+  if (a$rho1 != 0 || a$rho2 != 0) {
+    str1 = paste0(str1, ", FH(", a$rho1, ", ", a$rho2, ")")
+  }
+  
+  
+  
+  str2 <- paste0("Overall power: ", round(a$overallReject, 3))
+  
+  str3 <- paste0("Expected # events: ",
+                 round(a$expectedNumberOfEvents, 1))
+  
+  str4 <- paste0("Expected # dropouts: ",
+                 round(a$expectedNumberOfDropouts, 1))
+  
+  str5 <- paste0("Expected # subjects: ",
+                 round(a$expectedNumberOfSubjects, 1))
+  
+  str6 <- paste0("Expected study duration: ",
+                 round(a$expectedStudyDuration, 1))
+  
+  str7 <- paste0("Accrual duration: ",
+                 round(a$accrualDuration, 1), ", ",
+                 "fixed follow-up: ", a$fixedFollowup)
+  
+  df1 = data.frame(x = rep("", 8))
+  colnames(df1) = NULL
+  row.names(df1) = c(str1, str2, str3, str4, str5, str6, str7, "")
+  
+  
+  if (k>1) {
+    b <- data.frame(a$cumulativeRejection,
+                    a$cumulativeFutility,
+                    a$numberOfEvents,
+                    a$numberOfDropouts,
+                    a$numberOfSubjects,
+                    a$analysisTime)
+    
+    # format number of digits after decimal for each column
+    j1 <- c(3,4,5,6)
+    j3 <- c(1,2)
+    
+    b[j1] <- lapply(b[j1], formatC, format = "f", digits = 1)
+    b[j3] <- lapply(b[j3], formatC, format = "f", digits = 3)
+    
+    df = t(b)
+    rownames(df) = c("Cumulative rejection",
+                     "Cumulative futility",
+                     "Number of events",
+                     "Number of dropouts",
+                     "Number of subjects",
+                     "Analysis time")
+    
+    colnames(df) <- paste("Stage", seq_len(ncol(df)), sep=" ")
+  }
+  
+  print(df1, ..., na.print = "" , quote = FALSE )
+  
+  if (k>1) {
+    print(df, ..., na.print = "" , quote = FALSE )
+  }
+  
   invisible(x)
 }